# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Media AI API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/products/{id}/first-frame": {
      "get": {
        "tags": [
          "products"
        ],
        "summary": "GET products id first frame",
        "operationId": "getProductsIdFirstFrame",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ipId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FirstFrameList"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "products"
        ],
        "summary": "POST products id first frame",
        "operationId": "postProductsIdFirstFrame",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FirstFrameSaveResponse"
                }
              }
            }
          },
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FirstFrameSaveResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
...
    "/api/products/{id}/generate-video": {
      "post": {
        "tags": [
          "products"
        ],
        "summary": "POST products id generate video",
        "operationId": "postProductsIdGenerateVideo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoGenerationResponse"
                }
              }
            }
          },
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoGenerationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateVideoRequest"
              }
            }
          }
        }
      }
    },
...
    "/api/products/{id}/model-image/save": {
      "post": {
        "tags": [
          "products"
        ],
        "summary": "POST products id model image save",
        "operationId": "postProductsIdModelImageSave",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelImageSaveResponse"
                }
              }
            }
          },
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelImageSaveResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
...
    "/api/products/{id}/style-image/save": {
      "post": {
        "tags": [
          "products"
        ],
        "summary": "POST products id style image save",
        "operationId": "postProductsIdStyleImageSave",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StyleImageSaveResponse"
                }
              }
            }
          },
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StyleImageSaveResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GenerateVideoRequest": {
        "type": "object",
        "properties": {
          "ipId": {
            "type": "string"
          },
          "firstFrameUrl": {
            "type": "string"
          },
          "movementId": {
            "type": "string"
          },
          "productMaterialId": {
            "type": "string"
          },
          "step": {
            "type": "string"
          },
          "styleImageId": {
            "type": "string"
          },
          "sceneId": {
            "type": "string"
          },
          "composition": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "poseId": {
            "type": "string"
          },
          "prompt": {
            "type": "string"
          },
          "firstFrameId": {
            "type": "string"
          },
          "modelImageId": {
            "type": "string"
          }
        },
        "description": "Payload for first-frame generation and final video generation."
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {}
        },
        "required": [
          "error"
        ],
        "description": "Standard error response."
      },
...
      "ModelImageSaveResponse": {
        "type": "object",
        "properties": {
          "modelImageUrl": {
            "type": "string"
          },
          "modelImageId": {
            "type": "string"
          }
        },
        "required": [
          "modelImageUrl",
          "modelImageId"
        ],
        "description": "Saved model image response."
      },
      "StyleImageSaveResponse": {
        "type": "object",
        "properties": {
          "styledImageUrl": {
            "type": "string"
          },
          "styleImageId": {
            "type": "string"
          }
        },
        "required": [
          "styledImageUrl",
          "styleImageId"
        ],
        "description": "Saved styled image response."
      },
...
      "FirstFrameSaveResponse": {
        "type": "object",
        "properties": {
          "firstFrameUrl": {
            "type": "string"
          },
          "firstFrameId": {
            "type": "string"
          }
        },
        "required": [
          "firstFrameUrl",
          "firstFrameId"
        ],
        "description": "Saved first frame response."
      },
      "FirstFrameList": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string"
            },
            "productId": {
              "type": "string"
            },
            "ipId": {
              "type": "string"
            },
            "styleImageId": {
              "type": [
                "string",
                "null"
              ]
            },
            "url": {
              "type": "string"
            },
            "prompt": {
              "type": [
                "string",
                "null"
              ]
            },
            "sceneId": {
              "type": [
                "string",
                "null"
              ]
            },
            "composition": {
              "type": [
                "string",
                "null"
              ]
            },
            "inputHash": {
              "type": "string"
            },
            "createdAt": {
              "type": "string",
              "format": "date-time",
              "description": "ISO 8601 date-time string."
            }
          },
          "required": [
            "id",
            "productId",
            "ipId",
            "styleImageId",
            "url",
            "prompt",
            "sceneId",
            "composition",
            "inputHash",
            "createdAt"
          ],
          "description": "Generated first frame."
        }
      },
...
      "VideoGenerationResponse": {
        "type": "object",
        "properties": {
          "videoId": {
            "type": "string"
          },
          "videoUrl": {
            "type": "string"
          }
        },
        "required": [
          "videoId",
          "videoUrl"
        ],
        "description": "Generated video response."
      }
    },
...
  }
}